Move validation functions to verification file in device (#848)
**Context**
We have an empty `verification.py` file in the device module that was
intended as the long-term location of validation functions.

**Description of change**
Copy-pasted all the validation from `programs/verification.py` into
`device/verification.py`, and changed import paths
